+++ assets/javascripts/kitten/components/tooltips/static-tooltip.js
@@ -2,16 +2,32 @@ import React from 'react'
 import { Button } from 'kitten/components/buttons/button'
 
 export class StaticTooltip extends React.Component {
-  render() {
+  renderButton() {
     return (
-      <div className="k-StaticTooltip">
-        <Button className="k-Button k-Button--helium">
-          Button
-        </Button>
-        <div className="k-StaticTooltip__content">
-          Bonjour
+      <Button className="k-Button k-Button--helium">
+        Button
+      </Button>
+    )
+  }
+
+  render() {
+    const { id,
+            children,
+    } = this.props
+
+    return (
+      <div className="k-StaticTooltip"
+           id={ id }>
+        { this.renderButton() }
+        <div className="k-StaticTooltip__content"
+             id={ id }>
+          { children }
         </div>
       </div>
     )
   }
+}
+
+StaticTooltip.defaultProps = {
+  id: 'Toto',
 }