+++ assets/javascripts/kitten/components/tooltips/tooltip.js
@@ -1,82 +1,25 @@
 import React from 'react'
-import classNames from 'classnames'
-import { findDOMNode } from 'react-dom'
 import ReactTooltip from 'react-tooltip'
 import ButtonTooltipIcon from 'kitten/components/buttons/button-tooltip-icon'
 
 export class Tooltip extends React.Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      open: false,
-    }
-
-    this.handleClick = this.handleClick.bind(this)
-    this.show = this.show.bind(this)
-  }
-
-  componentDidMount() {
-    if (this.props.alwaysOpen) {
-      setTimeout(this.show, 1000)
-    }
-  }
-
-  handleClick() {
-    if (this.props.alwaysOpen)
-      return
-
-    if (this.state.open) {
-      this.hide()
-    } else {
-      this.show()
-    }
-  }
-
-  show() {
-    ReactTooltip.show(findDOMNode(this.button))
-    this.setState({ open: true })
-  }
-
-  hide() {
-    ReactTooltip.hide(findDOMNode(this.button))
-    this.setState({ open: false })
-  }
-
   render() {
-    const { className,
-            place,
-            alwaysOpen,
+    const { place,
             children,
-            refElement,
-            refElementProps,
-            tooltipModifier,
-            id } = this.props
-
-    const RefTag = refElement
-
-    const tooltipContainerClassName = classNames(
-      'k-Tooltip',
-      className,
-    )
-
-    const tooltipClassName = classNames(
-      'k-Tooltip__content',
-      `k-Tooltip__content--${tooltipModifier}`
-    )
+            id,
+            ...buttonTooltipIconProps } = this.props
 
     return (
-      <div className={ tooltipContainerClassName }>
-        <RefTag ref={ button => this.button = button }
-                   data-tip="tooltip"
-                   data-for={ id }
-                   onClick={ this.handleClick }
-                   aria-describedby={ id }
-                   { ...refElementProps } />
+      <div className="k-Tooltip">
+        <ButtonTooltipIcon data-tip
+                           data-for={ id }
+                           data-event="click"
+                           aria-describedby={ id }
+                           { ...buttonTooltipIconProps } />
         <ReactTooltip id={ id }
                       // This is not a mistake, this attribute is called
                       // class not className!
-                      class={ tooltipClassName }
+                      class="k-Tooltip__content"
                       role="tooltip"
                       effect="solid"
                       place={ place }
@@ -92,7 +35,4 @@ Tooltip.defaultProps = {
   id: '',
   place: 'right',
   children: null,
-  alwaysOpen: false,
-  tooltipModifier: 'hydrogen',
-  refElement: ButtonTooltipIcon,
 }